backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-multipart
psycopg2-binary
openai
python-dotenv
email-validator